requirements.txt
absl-py==1.3.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
audioread==3.0.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
comtypes==1.1.10
confluent-kafka==2.5.3
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.0
distlib==0.3.6
Django==4.2.6
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
djongo==1.3.6
dm-tree==0.1.8
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.3
filelock==3.6.0
Flask==2.1.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
flint==0.0.1
fonttools==4.29.1
fsspec==2024.3.0
future==0.18.3
galois==0.1.1
gast==0.4.0
google-api-core==2.17.1
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.25.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.13.0
h5py==3.10.0
idna==3.3
imbalanced-learn==0.10.1
imblearn==0.0
ipykernel==6.15.2
ipython==7.34.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.3.5
kafka==1.3.5
keras==3.0.5
kiwisolver==1.3.2
langdetect==1.0.9
lazy_loader==0.3
libclang==14.0.6
librosa==0.10.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mediapipe==0.9.1.0
ml-dtypes==0.3.2
mongoengine==0.24.2
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.8
Naked==0.1.31
namex==0.0.7
nbformat==5.8.0
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.7
nsepy==0.7
numba==0.56.2
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-contrib-python==4.7.0.72
opencv-python==4.5.5.62
openpyxl==3.1.2
openvino==2024.0.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.6.0
plotly==5.14.1
pooch==1.8.1
prompt-toolkit==3.0.23
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
pymatrix==3.0.1
pymongo==4.3.2
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
pyrsistent==0.19.3
PyScreeze==0.1.28
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-whois==0.8.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.2
pywin32==302
PyYAML==6.0
pyzmq==23.2.1
regex==2022.7.25
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scapy==2.5.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.2
selenium==4.1.3
selenium-requests==1.4.1
shellescape==3.8.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.1
soxr==0.3.7
SpeechRecognition==3.10.1
spyder-kernels==2.3.3
sqlparse==0.4.4
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
textblob==0.17.1
threadpoolctl==3.1.0
tldextract==3.2.1
tomli==2.0.1
torch==2.2.1
tornado==6.2
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.10.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.17.1
wcwidth==0.2.5
webdriver-manager==3.5.4
Werkzeug==2.1.2
wikipedia==1.4.0
windows-curses==2.3.1
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.1.0
xgboost==1.7.2
